+++ changelog.md
@@ -1,3 +1,19 @@
+# v3.16.1
+
+## Changes
+
+### Chat commands
+
+Added chat command to directly open the Master Looter options menu:
+
+`/rc ml`
+
+## Bugfixes
+
+- *Reduced voting frame row jumping.*
+- *Switching characters will no longer apply cached state.*
+- *Removing an item from the session frame mid session now actually removes it, i.e. the voting frame no longer thinks there's still items to award.*
+
 # v3.16.0
 
 ## Changes